healthcare/requirements.txt
google-generativeai==0.3.2
langchain==0.1.0
chromadb==0.4.22
python-dotenv==1.0.0
streamlit==1.24.0
pypdf>=3.0.0
pdfminer.six==20221105
tiktoken==0.5.2
python-magic==0.4.27
openai>=1.12.0
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0